Install modul and register command. Move all files to the modul.
