+++ Modules/PamyraOrder/composer.json
@@ -1,0 +1,31 @@
+{
+    "name": "nwidart/pamyraorder",
+    "description": "",
+    "authors": [
+        {
+            "name": "Nicolas Widart",
+            "email": "n.widart@gmail.com"
+        }
+    ],
+    "extra": {
+        "laravel": {
+            "providers": [],
+            "aliases": {
+
+            }
+        }
+    },
+    "autoload": {
+        "psr-4": {
+            "Modules\\PamyraOrder\\": "",
+            "Modules\\PamyraOrder\\App\\": "app/",
+            "Modules\\PamyraOrder\\Database\\Factories\\": "database/factories/",
+            "Modules\\PamyraOrder\\Database\\Seeders\\": "database/seeders/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Modules\\PamyraOrder\\Tests\\": "tests/"
+        }
+    }
+}
