+++ Modules/PamyraOrder/database/seeders/PamyraOrderDatabaseSeeder.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Modules\PamyraOrder\database\seeders;
+
+use Illuminate\Database\Seeder;
+
+class PamyraOrderDatabaseSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     */
+    public function run(): void
+    {
+        // $this->call([]);
+    }
+}
