+++ Modules/PamyraOrder/resources/views/index.blade.php
@@ -1,0 +1,7 @@
+@extends('pamyraorder::layouts.master')
+
+@section('content')
+    <h1>Hello World</h1>
+
+    <p>Module: {!! config('pamyraorder.name') !!}</p>
+@endsection
